pyproject.toml
[project]
name = "ali"
version = "0.1.0"
description = "Action Language Interpreter - One command for every tool"
readme = "README.md"
authors = [{ name = "Fredrik Angelsen", email = "fredrikangelsen@gmail.com" }]
requires-python = ">=3.12"
dependencies = ["lark>=1.2.2", "pyyaml>=6.0.2", "tmux-popup>=0.2.0"]

[project.scripts]
ali = "ali.cli:main"

[build-system]
requires = ["uv_build>=0.8.8,<0.9.0"]
build-backend = "uv_build"

[tool.basedpyright]
typeCheckingMode = "standard"

[tool.ruff.lint]
# Ignore star import warnings in language implementation files
extend-ignore = [
    "F403", # Star imports
    "F405", # Star import usage
]
